refactoring: removed the api package
