+++ src/main/java/com/github/lukaszbudnik/roxdb/application/RoxDBConfig.java
@@ -1,3 +1,3 @@
-package com.github.lukaszbudnik.roxdb.api;
+package com.github.lukaszbudnik.roxdb.application;
 
 public record RoxDBConfig(int port, String dbPath) {}
