+++ src/main/java/com/github/lukaszbudnik/roxdb/db/RoxDB.java
@@ -1,4 +1,4 @@
-package com.github.lukaszbudnik.roxdb.api;
+package com.github.lukaszbudnik.roxdb.db;
 
 import java.util.List;
 import java.util.Optional;
